Create SECURITY.md
or maybe- give them explicit instructions about what the hell is going on instead of letting their life blow apart. Provide detailed explanations to patients and staff of how their files or information could be breached and what can be done or is being done to prevent that. Maybe even adopting a plan from a noteworthy source. 
for example:
BSWNeurotexas complies with the EU-U.S. Privacy Shield Framework and/or the Swiss-U.S. Privacy Shield  Framework(s), as applicable) as set forth by the U.S. Department of Commerce regarding the collection, use, and retention of personal information transferred from both European Union and/or Switzerland, as applicable, to the United States.  BSWNeurotexas has certified to the Department of Commerce that it adheres to the Privacy Shield Principles.  If there is any conflict between the terms in this privacy policy and the Privacy Shield Principles, the Privacy Shield Principles shall govern.  To learn more about the Privacy Shield program, and to view our certification, please visit https://www.privacyshield.gov/.